SPEC/MGM: Bump dependency to XRootD 6 and fix API compatibility

+++ mgm/drain/DrainTransferJob.hh
@@ -64,8 +64,9 @@ public:
   //! @param source         the source url of the current job
   //! @param destination    the destination url of the current job
   //----------------------------------------------------------------------------
-  void BeginJob(uint16_t jobNum, uint16_t jobTotal, const XrdCl::URL* source,
-                const XrdCl::URL* destination) override
+  void
+  BeginJob(uint32_t jobNum, uint32_t jobTotal, const XrdCl::URL* source,
+           const XrdCl::URL* destination) override
   {
     using namespace std::chrono;
     mStartTimestampSec = duration_cast<seconds>
@@ -79,8 +80,8 @@ public:
   //! @param bytesProcessed bytes processed by the current job
   //! @param bytesTotal     total number of bytes to be processed by job
   //----------------------------------------------------------------------------
-  void JobProgress(uint16_t jobNum, uint64_t bytesProcessed,
-                   uint64_t bytesTotal) override
+  void
+  JobProgress(uint32_t jobNum, uint64_t bytesProcessed, uint64_t bytesTotal) override
   {
     mBytesTransferred = bytesProcessed;
     mProgress = static_cast<int>((1.0 - (1.0 * (bytesTotal - bytesProcessed) /
@@ -91,7 +92,8 @@ public:
   //! Determine whether the job should be canceled - this is used internally
   //! by the XrdCl::CopyProcess.
   //----------------------------------------------------------------------------
-  bool ShouldCancel(uint16_t jobNum) override
+  bool
+  ShouldCancel(uint32_t jobNum) override
   {
     return mDoCancel;
   }
